Split tensor product points and single point kernels
